chore: Organise entities into a directory (off-ticket) (#865)
